# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API davidWebProject",
        "description": "Cette API est conçue pour gérer l'intégration de diverses fonctionnalités dans mon projet de freelance. Elle permet la gestion complète du backend, incluant la création, la mise à jour, la suppression et la récupération de données essentielles telles que les témoignages clients, les projets, les utilisateurs, et d'autres éléments de contenu. L'API alimente mon site web en fournissant les données nécessaires tout en facilitant la gestion administrative des opérations. Elle est sécurisée via l'authentification par jeton et est optimisée pour une utilisation par des applications web et mobiles. Idéale pour un usage interne, cette API centralise et simplifie la gestion de mon activité freelance.",
        "contact": {
            "name": "Support",
            "url": "https://www.davidwebprojects.fr",
            "email": "contact@davidwebprojects.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://apifreelance.davidwebprojects.fr",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Lister tous les clients",
                "operationId": "getCustomers",
                "responses": {
                    "200": {
                        "description": "Liste des clients",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Créer un nouveau client",
                "operationId": "createCustomer",
                "requestBody": {
                    "description": "Données du client à créer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Client créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Afficher les détails d'un client",
                "operationId": "getCustomerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du client",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Mettre à jour un client existant",
                "operationId": "updateCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données du client à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Customer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Supprimer un client",
                "operationId": "deleteCustomer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du client à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Client supprimé avec succès"
                    },
                    "404": {
                        "description": "Client non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "title": "Customer",
                "description": "Modèle représentant un client",
                "required": [
                    "customer_number",
                    "status_id",
                    "created_by"
                ],
                "properties": {
                    "id": {
                        "description": "ID unique du client",
                        "type": "integer"
                    },
                    "customer_number": {
                        "description": "Numéro de référence unique du client",
                        "type": "string"
                    },
                    "civility_id": {
                        "description": "ID de la civilité associée au client",
                        "type": "integer",
                        "nullable": true
                    },
                    "company_name": {
                        "description": "Nom de la société du client",
                        "type": "string",
                        "nullable": true
                    },
                    "last_name": {
                        "description": "Nom de famille du client",
                        "type": "string"
                    },
                    "first_name": {
                        "description": "Prénom du client",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Numéro de téléphone du client",
                        "type": "string"
                    },
                    "email": {
                        "description": "Adresse email du client",
                        "type": "string",
                        "format": "email"
                    },
                    "address": {
                        "description": "Adresse du client",
                        "type": "string"
                    },
                    "postal_code": {
                        "description": "Code postal du client",
                        "type": "string"
                    },
                    "city": {
                        "description": "Ville du client",
                        "type": "string"
                    },
                    "country": {
                        "description": "Pays du client",
                        "type": "string",
                        "nullable": true
                    },
                    "additional_info": {
                        "description": "Informations supplémentaires concernant le client",
                        "type": "string",
                        "nullable": true
                    },
                    "status_id": {
                        "description": "ID du statut associé au client",
                        "type": "integer"
                    },
                    "user_id": {
                        "description": "ID de l'utilisateur associé au client",
                        "type": "integer",
                        "nullable": true
                    },
                    "created_by": {
                        "description": "ID de l'utilisateur qui a créé le client",
                        "type": "integer"
                    },
                    "updated_by": {
                        "description": "ID de l'utilisateur qui a mis à jour le client",
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Date et heure de création du client",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date et heure de la dernière mise à jour du client",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}